third-party/knowledge-catalog-main/samples/enrichment/src/requirements.txt
google-adk
google-auth
google-cloud-bigquery
google-cloud-dataplex
mcp[cli]
pyyaml
